[ci] fix table row highlighting when using on-page filter (#8818)
* [ci] fix table row highlighting when using on-page filter

The main issue is that `nth-child-of` still sees the hidden elements. We can instead
use `nth-of-type` and wrap hidden `tr`s in a `div`. Because the `div` is not a `tr`,
it does not appear as part of the `nth-of-type` calculation. The content is still
present on the page, so we can restore it by removing the `display: none` div.

* fix