[query] Replace recursive BTree foreach with iterative version
Allocate a fixed stack, since I couldn't find a good staged one that
grows, and this was simplest. 128 stack slots should be more than enough
for this implementation as it requires one stack slot per level of the
tree.

There are many improvements that can be made here, but hopefully this
should unblock some amount of the method splitting work.