[query] Replace recursive BTree foreach with iterative version (#8847)
* [query] Replace recursive BTree foreach with iterative version

Allocate a fixed stack, since I couldn't find a good staged one that
grows, and this was simplest. 128 stack slots should be more than enough
for this implementation as it requires one stack slot per level of the
tree.

There are many improvements that can be made here, but hopefully this
should unblock some amount of the method splitting work.

* Fix scoping bug in CollectAsSet and Grouped aggs

We need to be sure to put locals on the right method. Using the
StagedRegionValueBuilder's MethodBuilder leads to locals being put on
the wrong method.

* Fully inline AppendOnlyBTree foreach

I kept the method generation because I could not figure out why the
locals were being placed on the wrong method. Having fixed that, we can
now inline foreach.