[dndarray] Implement hl.experimental.dnd.DNDArray (#8864)
* [dndarray] Implement hl.experimental.dnd.DNDArray

I plan to move this to `hl.dnd.DNDArray`.

I had to make a couple changes to RVD and Table to make this work. They all
revolve around convincing Hail not to elide *critical* `key_by`s.

The critical insight is that 1:1 partitioners (partitioners where each range
bound interval contains exactly one key) are special: permuting their keys is
free.

I can take advantage of this by combining two changes:
1. `RVD.enforceKey` is aware of these partitioners and avoids scans in that case
2. Defeat the optimizer, which is unaware of these partitioners and misoptimizes
   to operations that require shuffles.

The first change is easy. I added `RVDPartitioner.keysIfOneToOne` which looks
for these kinds of partitioners in the special case of keys consisting of 32-
and 64-bit integers.

The second change eluded me for a long time. Finally, I discovered
`isSorted=true` and realized the optimizer refuses to modify such
`TableKeyBy`s. I exposed this field in Python as: `Table._key_by_assert_sorted`.

With this infrastructure in place, I was able to implement read, write, and
matrix-multiply for DNDArray!

In addition, to the arguable hacks above, a couple pain points remain:
1. I do not know how to rename keys in Python without triggering shuffles. If I
   write `key_by(x=t.y, y=t.x)`, Hail implements this as
   `TableKeyBy(TableMapRows(TableKeyBy(Array(), ...)`. The inner key by throws
   the keys away so that they can be modified with TableMapRows. Unfortunately,
   this completely defeats my attempts to avoid shuffles. I avoid this issue by
   not using fixed names for the x and y block coordinates (their names are
   stored in `x_field` and `y_field`).
2. Hail lacks `ndarray_sum`. Instead, I convert from ndarray to array so that I
   can use `array_sum`. Unfortunately, this operation seems to completely
   dominate all of my time. It takes about 10x as much time as the matrix
   multiplies take. I do not understand this. I should be reading the entries in
   column-major order.

Performance
-----------

```
In [1]: %%time
   ...: import hail as hl
   ...: mt = hl.balding_nichols_model(n_populations=2,
   ...:                               n_variants=10000,
   ...:                               n_samples=10000,
   ...:                               n_partitions=100)
   ...: mt = mt.select_entries(gt = hl.float(mt.GT.n_alt_alleles()))
   ...: da = hl.experimental.dnd.array(mt, 'gt')
   ...: da.write('/tmp/in.da', overwrite=True)
In [3]: %%time
   ...: bm = hl.linalg.BlockMatrix.from_entry_expr(mt.gt)
In [5]: %%time
   ...: (bm @ bm.T).write('/tmp/foo.bm', overwrite=True)
In [7]: %%time
   ...: import hail as hl
   ...: da = hl.experimental.dnd.read('/tmp/in.da')
   ...: (da @ da.T).write('/tmp/out.da', overwrite=True)
```

Block matrix performed the matrix multiply in 19.3s. DNDArray performed the
matrix multiply in 37.6s.

* relax privacy

* addressed

* wip

* wip

* fix

* kbye

* kbye

* kbye

* kbye

* kbye

* fix

* fix

* fixes

* fix

* handle missingness

* fix

* fix

* fixes

* renames

* fixg

* fix

* fix type