[ci] make ci idempotent, try two (#8833)
Database idempotency is easy, use `IF NOT EXISTS`. User cleanup is hard because
the job isn't deterministic. Each attempt will produce a different password. You
cannot retrieve a password of an extant user.

The bulk of the changes are in service of creating a job whose attempts will
race to create the password. Only one attempt will win that race and pass its
password on to the `create_database_job`. Those jobs will race to create
databases and users and secrets, but all those operations are now
idempotent (`CREATE USER` uses `IF NOT EXISTS`; the secret is always the same).