[ci] fix clone to actually retry
I do not understand how `set -e` works in bash functions, but I know what
currently exists does not retry properly. I have verified this works as intended.

It deletes *all* files in the current folder, then tries to clone into this folder.
If the clone fails, becasue it is the last command in the function, the function
will return the same exit code. The retry function will then retry `clone`. Because
`clone` removes all files in the directory before attempting a clone, it is idempotent.

Users of the checkout script all use it by creating an empty directory and cd'ing into it.