[query] Fix memory leak in combOpFSerialized (#9028)
The previous implementation, while seeming to be well-abstracted at
first, actually had a rather devious property of creating agg states
for multiple classes multiple times. I'm still working on figuring out
*exactly* the place where our assumptions broke down, but this change
definitely fixes the problem, and simplifies the implementation by
directly using IR, instead of other compiled functions.

This problem was a symptom of a larger issue, which is that the
ownership semantics of the current aggregator system is way too complex
to be coding against regularly. This all will go away when lowering is
complete, in favor of the *much* simpler set of IR nodes that are used
in lowering. We may need to address this problem sooner, though.

CHANGELOG: Fixed memory leak affecting `Table.annotate` with scans, `hl.experimental.densify`, and `Table.group_by` / `aggregate`.