[query] Fix PCanonicalIntervalValue loadStart and loadEnd
Struct's loadField has been changed to load the address directly in the
case of the field being a collection type. PCanonicalIntervalValue was
using loadStart and loadEnd rather than startOffset and endOffset which
caused addresses to be loaded twice in the case of intervals of
collections, leading to segfaults.

This code is currently unused, but will be used in EmitCodeOrdering