[shuffler] IR (#8726)
* [hail] error if IR is defined in neither emit nor emitI

* [hail] teach begin to return its last value

* wip

* [hail][ir] improve Stream.unfold signature

- Teach `unfold` to accept close and close0.
- Teach `unfold` to use the language of `Stream`, i.e. `setup0` and `setup`.
- Teach `unfold` to not require `setup` and `setup0`.

* [hail][ir] teach function registry to raise helpful errors

When you screw up an `Apply` IR, you previously received no information as
to why. This change teaches the function registry to display all signatures
that match the function name but not the types, assisting a developer in
diagnosing whether they got the name wrong or the types wrong.

* [hail][ir] clean up IRFunctionRegistry

I rage programmed here a bit. I was pretty frustrated with this class when
trying to debug some other issues in the Shuffler IR. I have two aims with this
PR:
1. Use a clear, consistent naming scheme throughout the file
2. Use concise definitions

In particular, I unified the terminology around types to use these words:
- valueParameterTypes: the type of each value-level parameter
- typeParameters: the type-level parameters, these are always type variables, I
  considered calling them typeVariables, but I like the symmetry with valueParameters
- returnType and returnPType
- typeArguments and valueArguments

I also simplified some anonymous class definitions by providing constructor arguments
instead of methods that are always overridden by `val`s.

Unfortunately, when I got to compiling the tests I discovered we create a function type
that does not exist anywhere in the real code. This did not fit my scheme well and lead
to the duplicative definition of SeededMissingAwareJVMFunction.

Oh, and I changed `IRFunction` to `JVMFunction` because there are already "IR" functions
and an "irRegistry" and it was super confusing to not have the IRFunctions inside the
"irRegsitry". I did not use `CodeFunction` because these are actually implemented by
a number of different JVM Bytecode building tools: `Code`, `PCode`, `EmitCode`, and
`IEmitCode`.

* wip

* [hail][etype] teach EType how to make OutputBuffer => Encoder

* refactor

* [hail][ptype] improve assertion message

* [hail][io] OutputBlockBuffer *must* also flush

Anything with an `OutputStream` must be able to `flush`
said `OutputStream`.

* [hail][io] Teach TypedCodecSpec to create decoders without a VType

A TypedCodecSpec has a reasonable VType on hand: `_vType`. In the Shuffler, I have
found myself repeating myself too often. This seems generally useful.

* [hail][io] implement read/write UTF

The Shuffler uses this

* fix

* fixes

* fixes

* fix

* [hail][io] tracing buffers and streams for debugging

These are unused, and should mostly remain that way, but it is convenient to
have them on-hand when you need to figure out why an Encoder or Decoder is
misbehaving. A `TracingOutputBuffer` is missing because I did not need it
and have not spent the time to write it.

* fix

* fix

* fix

* [hail][asm4s] implement _assert

* [hail][ir] improve error message when unify fails in PruneDeadFields

With this added context it was completely obvious what I had done wrong. Without
this context, I just knew that somewhere I had some ints and structs that did not
unify.

* [hail] teach MemoryBuffer to act like ByteArray(In|Out)putStream

The Shuffler uses MemoryBuffer to apply Hail encoders and decoders on byte arrays. For
example, I need to serialize the ETypes into generated bytecode so I can send it to the
Shuffler at run-time. I serialize the EType object to bytes using an encoder and MemoryBuffer
Then, I base64 encode it and store it as a JVM constant pool String. When I need to start
a Shuffle I base64 decode the string, stuff the bytes in a MemoryBuffer and decode it.

* fix formatting

* [hail][ir] assert EmitStream.unfold f is not null

* address comments

* [hail][io] OutputBufferValue

* [hail][io] InputBufferValue

* fixes

* Revert "address comments"

This reverts commit fe6f59ee431d30d26223a109fe7bcfaa36d8cee0.

* Revert "fix"

This reverts commit a480de6bc717c0cc76382a143598e8e2815db8c1.

* Revert "fix"

This reverts commit 698bba28559a8ddb7432c3643b84faadaa2b4d09.

* Revert "wip"

This reverts commit e6a52b23febf332cab75ccabc194dfe8b43c092a.

* Revert "[hail] teach begin to return its last value"

This reverts commit 3bb63fce85ce60db38a1a02ae7f7859e70f93989.

* fix

* [hail][internal] Improve error messages when normalize names fails

In particular, include both the IR context (a series of IR names) and the
environment context (names to iruids). This enables me, as a developer, to
rapidly identify the problematic node.

* [hail] refactor RPrimitive.typeSupported

I also converted a no-message assertion into an if with an AssertionError
because I found it difficult to debug without the added information.

This removes the duplication of the list of supported types for RPrimitive.

* [hail][internal] teach EType to preserve XStruct

This mirrors the functionality available on `TypedCodecSpec`. In some cases (the shuffler),
you might be handed a buffer that is already configured, but you still want to create a
decoder whose PType is known to be a subtype of PStruct.

* [hail][internal] Teach TypedCodecSpec to supply decodedPType

Just as `buildEmitDecoderF` need not accept a virtual type, we might want to know
the PType that the natural decoder would produce. This PR implements that.

* works

* wip

* wip

* fix

* fixes

* sorted=false

* fix

* fixes

* [hail][sbt] bloop 1.4.0 has a bug that breaks Java 8 compatibility

Fix is to upgrade to Bloop 1.4.1. https://github.com/scalacenter/bloop/issues/1276

* [shuffler] incremental improvements

These incremntal improvements set us up for a cleaner diff when the Shuffler IR lands.

* fix

* fix?

* fix

* address comments

* [ci][build] do not fail if database already exists

Five things may already exist:
- creating the database
- creating the admin or user user
- creating the admin or user secret

Creating the database is now idempotent with `IF NOT EXISTS`.

Creating a user is idempotent because we create a user with a random name. If
we're racing with someone else, the database might have extra users
created.

Finally, we race to create the secrets. Whoever runs last wins the race and
their secret create is the one downstream tasks see. Secret creation was made
idempotent by using `create --dry-run` piped to `apply`. Apply does not fail if
the secret already exists, it merely updates it. In a sense, apply is an atomic
create-or-update.

* fixes

* more fixes

* huh

* fixes

fixg

wip

wip

wip

fix typecheck for  shufflewritingfinished

wip

fix keyRange type

wip

implement

wip

wip

wip

qip

whatever

iwip

wip

wip

fix

wip

wip

wip

broke?

broke?

wip

fix

fix

fix

fix

adapt to new unfold

fix logging

remove traces

* fixes

* fix

* fixg

* fixes

* fixes?

* wip

* fix

* fix

* fix

* fixes

* tls

* move to services

* wip

* wip

* wip

* wip

* wip

* remove keys

* revert

* revert

* revert

* retores

* fix

* fix

* fix

* wip

* fix requiredness

* fixes

* use shuffleclient in shufflewith and shufflewrite

* fix

* fix

* fix

* maybe fix the field issue

* also need to fix IRParser

* also needed to craete the SortOrder parse functionos

* fixes

* error on non-realizable types in shufflewith

* fix

* remove unuse code

* evict withCleanup

* rewrite ShuffleWith emit to not use withCleanup

* fix withCleanup changes again

* revert unused Region.pretty

* add blank line after method

* cleanup emit a bit, leverage new consume, remove dead code

* remove a use of handle in favor of consume

* tighten shuffle uuid code

* simpler

* dead code

* fix compile errors

* fix fields

* remove extraneous shit

* undo buffer stuff

* revert x

* ir suite and cb.store

* pcanonicalbinary stuff

* fuckk

* works

* maybe address parsing issues

* use indexedseq for sensible equality