[notebook] fix notebook, improve tls naming, add typing to tls.py
The fix for Notebook is on line 242. Copying from my Zulip post:

> I made a mistake when I implemented TLS.
>
> In the following code snippet we use ssl_client_session which should probably be
> called in_cluster_ssl_client_session. It's supposed to be used to communicate
> with other services in the cluster. That needs to be changed back to
> aiohttp.ClientSession which loads the normal system certificates (including the
> VeriSign root certs that signed the public certs that gateway uses, different
> from the internal certs that our services use).
>
> In particular, note that the error says "unable to get local issuer
> certificate." That means that the local trust store lacks a certificate that
> trusts the remote server's certificate. In Dania's case, the default python on
> OS X lacks all certificates, so every remote server is untrusted. In notebook's
> case, ssl_client_session creates an SSL/TLS session that only trusts Hail
> internal services (in particular, it does not trust the certificates that
> gateway uses for incoming public traffic). The error also says that the server
> in question is workshop.hail.is which is a public domain (note the hail.is), so
> that traffic is going through the public gateway with its public certificates.
>
> ```
> # don't have dev credentials to connect through internal.hail.is
> ready_url = deploy_config.external_url(
>     service,
>     f'/instance/{notebook["notebook_token"]}/?token={notebook["jupyter_token"]}')
> try:
>     async with ssl_client_session(
>             timeout=aiohttp.ClientTimeout(total=1),
>             headers=headers,
>             cookies=cookies) as session:
>         async with session.get(ready_url) as resp:
> ```

I also changed the names and functionality of the functions in tls. Now
`in_cluster_ssl_context` will error if there is no ssl configuration found
instead of silently (and confusingly) using an SSLContext suited for public
communication (and wrong for in-cluster communication).

I added `get_context_specific_client_ssl_context` which should only be used in
publicly consumable tools (*never* in a service). This function allows the same
tool to be used inside and outside the cluster. It will load the correct certs
for your environment (it will load public certs if you're outside the cluster,
it will load in-cluster-only certs if you're in the cluster).

I also added types to `tls.py` and fixed some type errors.