[address] implement the address service (#9129)
* [address] implement the address service

This is a name to IP address and port service. GKE exposes pod IPs onto our VDC
network. As such, regular Google Cloud VMs can access pods by IP. GKE cannot
expose our services as IPs on our VDC because the way services load balance
traffic is more complex than DNS can handle. We acknowledge and accept the
limitations of client-side load balancing. In particular, if there are not many
clients and clients re-use address-port-pairs traffic will likely be
unbalanced. This is not a problem for the planned Shuffle service because the
clients are intended to be numerous (consider all the workers in a Query or
Batch pipeline).

The big change is that deploy config now has an `addresses` function which will
return a list of address-port pairs. Deploy config also now has `address` which
randomly chooses one of the address-port pairs.

I have included a simple test.

* update router

* add address k8s Service

* fix makefile

* fix comments

* boom

* fixes

* fix

* fix

* test_address depends on create_accounts

* fix arg

* fix