[query/combiner] Add options to better partition imported GVCFs (#9224)
* [query/combiner] Add options to better partition imported GVCFs

Teach the VCF combiner three new options:
  * --genomes/-G to indicate that it is operating on genomes
  * --exomes/-E to indicate that it is operating on exomes
  * --import-interval-size to more finely control the partitioning of
    the reference genome on GVCF import.

These are specified as a required mutually exclusive group, exactly one
of them must be present.

This goes along with removing `default_exome_intervals` and adding
`calculate_even_genome_partitioning`. The new function takes a reference
genome and a target interval size and computes a partitioning of the
reference into intervals of length at most the target size. It does so
by looping over the 'regular' contigs of the reference (as such it only
supports GRCh37 and GRCh38 for now) and dividing each contig's length
by the target size and rounding up, obtaining the number of partitions
the contig will be subdivided into. Then dividing the length of the
contig by the number of partitions, getting the true partition size.
Then using the true partition size, generating the intervals. This
hopefully leads to few degenerate cases where some partitions at the end
of contigs are very small.

* fix tests