Remove assertion in TabixLineIterator.next
This assertion is a sanity check to make sure that if the reader reads
to the end of a line, it does so exactly. We don't position the reader
exactly, so the assertion is invalid in our case. We also don't care, if
the offset is a little past the end of a line (as it will basically only
happen for empty blocks). We're always going to either stop reading or
seek at this point.