[query] Fix corner case assertion failure in tabix read (#9304)
* [query] Fix corner case assertion failure in tabix read

The tabix line iterator works by building a list of virtual offsets from
the index that correspond to a requested interval. We recently
discovered an issue that would lead to a runtime exception if the
following conditions held:

* An offset pair ended exactly on a block boundary.
* The block boundary was exactly the start of a line.

In a blocked file with virtual offsets, there are two ways to point to
the start of every block, (previous block start offset, previous block size) or
(current block start offset, 0).

Because of the way curOff is calculated in TabixLineIterator, this would
lead to a situation where curOff was the (previous block start offset, previous block size)
value, causing the jump to next chunk comparision `!TbiOrd.less64(curOff, offsets(i)._2)`
to fail when it should succeed, causing an extra line to be read, which
then makes the assertion check in the next iteration.

The key invariant of TabixLineIterator must be to keep bufferCursor
inside the block currently being read. We therefore make a check to see
if we need to refresh the buffer in any circumstance where we read
a line. Coupled with a reflow of control in readLine to increase
readability and fix a bug where lines greater than 64k would not be read
properly.

* Add long line gvcf import test

I used something a little more interesting than just a repeated
character. Unzip the file if you want to see :P

* Remove assertion in TabixLineIterator.next

This assertion is a sanity check to make sure that if the reader reads
to the end of a line, it does so exactly. We don't position the reader
exactly, so the assertion is invalid in our case. We also don't care, if
the offset is a little past the end of a line (as it will basically only
happen for empty blocks). We're always going to either stop reading or
seek at this point.