[query] Switch _buildEncoder methods to use CodeBuilder infrastructure (#9371)
* Rename ClassBuilder variable from cb to kb

This is to not confict with CodeBuilder

* Change _buildEncoder interface to take a CodeBuilder

NOTE: broken

* EArray buildEncoder to CodeBuilder

* EPackedIntArray buildEncoder to CodeBuilder

Need to update this to be more amenable to PCode/PValue. Issue is
IntPacker uses the address directly in order to save memcpys.

* Primitive Types buildEncoder to CodeBuilder

* EBinary buildEncoder to codeBuilder

Eventual FIXME, use PBinaryValue, issue is, like EPackedArray, we need
the address directly for OutputBuffer.

* ENDArrayColumnMajor buildEncoder to codeBuilder

* EBlockMatrixNDArray buildEncoder to CodeBuilder

* EBaseStruct buildEncoder to CodeBuilder

* ETransposedArrayOfStructs buildEncoder to CodeBuilder

* Restore previous NDArray encoder loop codegen

Can't figure out what's wrong, but this gives correct answers.

* Reduce method name size

I was seeing ooms for very wide tables, probably having something to do
with method splitting copying these strings over and over again, for
now, reduce a method name's length.

* Reduce methodName max size to 2000 characters