Changed db.py to use dictionaries instead of formatted urls
For retrieving the correct url based on the user's region. DB.DB__by_name dictionary will only contain datasets that are available in a region (e.g. gnomad datasets not available in 'eu' yet). When creating Dataset object with Dataset.from_name_and_json(), DatasetVersion.get_region() will only grab the versions that are available in the selected region. If a version is not available for a region, version.check_region() raises a warning, and the dataset will be excluded from DB.DB__by_name when initializing DB object. Additional check in DB.annotate_rows_db() that raises a ValueError if user still tries to select a dataset that is not available.