[query] Fix data race in FASTA reader (#9427)
In readBlock, membership was checked in a LinkedHashMap, then if it was
found, the value was retrieved from the cache. This lead to a data race
where a value could be evicted between the check and retrival. The
solution is to grab the block value out of the map, and if it is not
null, return it, otherwise, grab the block string, put it in the map,
and return it.

Co-authored-by: Tim Poterba <tpoterba@broadinstitute.org>

Co-authored-by: Tim Poterba <tpoterba@broadinstitute.org>