[query] Work around concurrency issues in FASTAReader
Introduce FASTAReaderConfig to act as a kind of factory for FASTAReader,
while all FASTAReaders themselves are confined to ThreadLocals (except
in tests).

Furthermore, add a lock around the fasta file map to prevent more than
one fasta from being copied per jvm. The can be lock contention on the
map, but if there is large amounts of waiting for said lock, then it
usually means that a fasta is downloading and we definitely should be
waiting.