[batch] do not explicitly set cpu in BPE, which improves test time
The `BatchPoolExecutor` assumes, incorrectly, that the default number of CPUs for
a batch job is 1. This is not true in test and development environment.

I avoid explicitly setting the value if it is `None`. This choice preserves the
default value in this environment. I set the thread limit to 1 if the CPU is `None`.