Add TabixVCF, tabix index creation for VCFs
An 80% solution to the long requested tabix a VCF on export feature.
It currently requires a file to be fully written before indexing can
occur.

Use htsjdk's index builder on top of our BGzip readers/writers. It does
no parsing/verification, other than the second column must be an
integer. If a line doen't start with '#' the first tab separated column
will be used as the contig and the second will be used as the position.
It's pretty fragile and does not gracefully handle positions being not
integers or the file not having chr/pos in the first two columns. Error
reporting is basically nonexistant.