Add tabix writing for parallel output.
There is still an export stage and a tabix stage, but tabix will never
error as unsupported.

The intent of ExportType.PARALLEL_COMPOSABLE is to be used with gsutil
compose, it's largely undocumented at this point. It is not worth doing
extra work when the user has explicitly stated the intent to use an
external tool to combine the files. A warning is printed instead.