[query] Quick and dirty tabix creation/writing for export_vcf (#9485)
* [query] Encapsulate TabixLineIterator buffering logic

Split it out into BGZipLineReader.

* Add TabixVCF, tabix index creation for VCFs

An 80% solution to the long requested tabix a VCF on export feature.
It currently requires a file to be fully written before indexing can
occur.

Use htsjdk's index builder on top of our BGzip readers/writers. It does
no parsing/verification, other than the second column must be an
integer. If a line doen't start with '#' the first tab separated column
will be used as the contig and the second will be used as the position.
It's pretty fragile and does not gracefully handle positions being not
integers or the file not having chr/pos in the first two columns. Error
reporting is basically nonexistant.

* Add python API support for tabix on vcf export

It's currently only supported for ExportType.CONCATENATED, but should be
easily extendable to the other forms.

* Fix tests

Had an NPE from attempting to use the BGZ line. Forgot to update python
IR tests for new argument.

* Add tabix writing for parallel output.

There is still an export stage and a tabix stage, but tabix will never
error as unsupported.

The intent of ExportType.PARALLEL_COMPOSABLE is to be used with gsutil
compose, it's largely undocumented at this point. It is not worth doing
extra work when the user has explicitly stated the intent to use an
external tool to combine the files. A warning is printed instead.

* Address Comments

- Better message for parallel tabix
- Don't return null from the parallelize closure, could be bad for the
  service backend.
- Remove catchall branch, it's already validated and throwing a match
  error is fine.