[batch][auth][ci][monitoring][notebook] check for sql mutation
SQL migrations are not permitted to be modified. Unfortunately, our tests
previously did not verify this at all. Indeed, a PR merged which modified a SQL
file. This PR caused main to fail a deploy.

This change verifies that no SQL migration is mutated in the source SHA relative
to the target SHA. One can also use it locally by running `make
check-services` from the root. Unfortunately, it does not work properly when run
on the main branch because there is no obvious point of comparison.

I considered comparing against the previous commit, but that might cause
failures if we have to manually fix something in batch. As such, I prefer a
non-deploy only test.