[batch] Update flock strategy (#9523)
flock(2) does not prevent other processes from doing anything to the
locked file unless they also gate usages on an flock call. Update Flock
to open only one descriptor, the directory that is the parent of `path`,
creating said directory if it doesn't exist.

The semantics have changed slightly. Previously if path was an already
extant directory, it would be opened and that descriptor would be
flocked. Now, the parent is always locked.