[query] Expand Python version compatibility with spark 3
Spark 3 resolved the issues with library changes python 3.8, so we can
now use python 3.8 as long as pyspark is version 3 or greater.

I also took this opertunity to make every link in or welcome doc message
a canonical https link that will be recognized as such by many parsers
such as sufficiently advanced terminal emulators.