[batch] Teach BatchPoolExecutor how to handle container errors (#9543)
* [batch] Teach BatchPoolExecutor how to handle container errors

CHANGELOG: BatchPoolExecutor now raises an informative error message for a variety of "system" errors, such as missing container images.

If the main container fails for reasons beyond BatchPoolExecutor's control, such
as a missing container image, we previously did not report these errors. In
fact, we encountered errors when trying to load the output file that cannot
exist if the main container errors.

Smaller included changes:
- directly use the asynchronous, low-level client instead of the synchronous,
  low-level client
- introduce an `async_cancel` now that we have access to the async client.

* make cancel sync