[query] Add PCodeParam and PCodeParamType (#9595)
* [query] Add PCodeParam and PCodeParamType

Remove implicit conversion from ptype to EmitParamType, adding asParam
and asEmitParam to PType for explicit conversion.

Currently PCodeParams are implemented pretty poorly as we should really
be creating the proper PValue from the parameters. This will require
either expanding nCodes for PCode or automatically memoizing the value
when getPCodeParam is called.

* Add needed import