[hail][docs][site] XSLT is dead. Long live Jinja2!
High level take-aways:

- Hail docs now have syntax highlighting (we just needed to import pygments.css).
- Search works again.
- There are now only two root HTML templates: `site/templates/base.html` and
  `web_common/web_common/templates/layout.html`. I cannot unify these further
  because our services and our main websites actually differ significantly.
- The search/nav bar is now present in the HTML, no JS nonsense to
  asynchronously load it into place after HTML rendering.
- Site now has a `make watch` which watches for changes and automatically
  re-renders the HTML.
- Site now has a few make rules that facilitate experimenting with how the docs
  are displayed within the context of the current development version of site's
  CSS & HTML.
- XSLT is now only used by the C++ tests.

Smaller things:

- Removed bootstrap dependencies. Did we ever actually use these?
- Removed "clipboard.js" dependency. Also not clear from where this came.
- Removed use of the `subtitle` tag, which isn't actually an HTML tag?

Future work:

- Simplify our CSS. It's not possible to logically reason about our CSS. And it
  interacts in bad ways with the latent RTD themes. I want a unified Hail visual
  theme.
- Clean up the search-related JavaScript in nav-bottom.html and
  search.html. These both seem too complicated to just make search work.

---

The thrust of this PR is to restructure Hail's website and documentation to
entirely rely on Jinja2 templates. Previously, we used a mix of Jinja2, XSLT,
and in-browser JavaScript DOM manipulation to piece together a web page.

Now, all of Hail's non-service website derives from
`site/templates/base.html`. It is a Jinja2 template with four blocks: title,
meta_description, head, and content. It ensures that:
- Hail's CSS is loaded,
- the Hail icon is set,
- the fonts are loaded,
- the source code highlighter is loaded (prism.js, only used outside the docs)
- the nav bar is present and configured.

The nav bar is somewhat complicated. There are two pieces. `nav-top.html`
contains the nav bar HTML elements. `nav-bottom.html` contains the JavaScript
code that hooks the search bar up to Algolia and sets the active page in the
navigation. I believe JavaScript which modifies the HTML DOM is traditionally
placed at the bottom of the `body` tag so that it is executed *after* the HTML
DOM is mostly rendered. That's why the navigation/search bar is split across two
files. I also load the `prism.js` source code highlighter at the end of the
body.

All of the non-docs pages are defined by html files in `pages`. Each one of
these is a Jinja2 template which derives from the base template. `make render`
converts every template in `pages` into a real HTML file in `www`. Check out
404.html for a simple example.

Once I had the site in working order, I turned my eyes to the docs. I converted
`docs/_templates/layout.html`, the base template for our docs, into a template
which derives from `site/templates/base.html`. That ensures everyone is using
the same CSS, the same navigation/search bar, same icon set, etc. Convincing
Sphinx to work like this was actually really easy because Sphinx already uses
Jinja2 templates! I just added site's templates folder to the Sphinx
`templates_path`.

I eliminated a few conditionals that are only relevant if your docs are also
rendered on RTD's server, which ours are not.

Finally, in order to experiment quickly with this, I changed site's Makefile to
add some rules that keep the automatically render the site and docs in response
to file edits.