[services] pervasively gracefully shutdown
`asyncio.ensure_future` creates background tasks. When a service is shutdown by
kubernetes, these background tasks keep the service up for the full 30 second
grace-period. This impedes development and deployment velocity.

This change eliminates all uses of `asyncio.ensure_future` which ignore the
returned future. Instead, all these futures are managed by a
`aiotools.BackgroundTaskManager`. This is a new class which holds a (weak)
reference to all of its created tasks and provides `shutdown` to cancel all the
tasks. Canceling the tasks ensures a speedy shutdown.

The use of a weak reference is critical to enable the garbage collector to clean
up tasks that have completed successfully. In particular, the batch worker
creates short-lived, background tasks.