[ci] delete open batches rather than cancelling them
We cancel in two cases:
1. A (dev|) deploy failed.
2. A batch was found open but unknown to us.

In the former case, we should delete the batch. The batch contains no extra information
everything we need to know is in the failure message.

In the latter case, an open batch does not cost us anything and
cannot be cancelled anyway, so we choose to ignore them. Such a batch
should only come into existence when CI dies in the middle of submission
which should be quite rare.