[query/stypes] Introduce SCode
This commit introduces the SCode hierarchy. How are SCodes different
from PCodes? SCodes are what we want PCodes to be. They don't have
the methods `code / tcode`, which gives us a way to implement new
performance-improving types like SStackStruct with well-defined
boundaries around that functionality.

Currently, the `asPCode` and `IEmitCode.typecast` methods break this
boundary, and I added these as a short-term mechanism to avoid another
very spicy meatball.

This commit is entirely reorganizational, with no semantic changes
to the compiler.