[ci] fix two longstanding bugs (#9726)
* [ci] fix two longstanding bugs

Since the first build.py we have had two bugs. Dev deploys would regularly
re-deploy the router service in the default namespace. The more recent bug
is that the router port ought to be 443, not 80. Any time we dev deploy,
I suppose we have broken the mainline, but this must have been resolved by
deploys that came shortly thereafter.

* fix