[sync] alternative to ksync (#9759)
* [sync] alternative to ksync

I got fed up with ksync not working the way I wanted and wrote my own
version. Assuming `repo/devbin` is on your path, then:

```
sync.sh batch dking
```

will keep the web_common, gear, hailtop, and batch repositories all up to date
in k8s. It just copies all the files over from your local machine, so its not
the fastest thing in the world. It takes maybe 2 seconds to get back to a
working container.

It also assumes there's one container per-pod.

It is completely disabled in the default namespace (I remove it in
deployment.yaml if we are in deploy mode).

* wip py rewrite

* fixes

* fixes

* fxes

* fix

* fix

* fix permissions

* hopefully fix last bugs

* Update tasks.py

* fix

* fix the deployment