[hail] Add two new forms to any all
CHANGELOG: `hl.any` and `hl.all` now also support a single collection argument and a varargs of Boolean expressions

Any and all currently take two arguments: a function and a collection. This
differs from the Python `any` and `all` which take only a collection and check
for truthy values.

We rectify this situation by supporting three forms:

1. ``hl.all(boolean, ...)``, are all arguments ``True``?
2. ``hl.all(collection)``, are all elements of the collection ``True``?
3. ``hl.all(function, collection)``, does ``function`` return ``True`` for
    all values in this collection?